Made changes (pull request #13)
