--- Netflix/Extensions/UIExtensions.swift
@@ -1,34 +1,0 @@
-import UIKit
-
-// Set paddings fot UITextField
-extension UITextField {
-    func setLeftPadding(_ amount: CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
-        self.leftView = paddingView
-        self.leftViewMode = .always
-    }
-    func setRightPadding(_ amount: CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
-        self.rightView = paddingView
-        self.rightViewMode = .always
-    }
-}
-
-// Set scroll to for UIScrollView
-extension UIScrollView {
-    func scrollTo(horizontalPage: Int? = 0, verticalPage: Int? = 0, animated: Bool? = true) {
-        var frame: CGRect = self.frame
-        frame.origin.x = frame.size.width * CGFloat(horizontalPage ?? 0)
-        frame.origin.y = frame.size.width * CGFloat(verticalPage ?? 0)
-        self.scrollRectToVisible(frame, animated: animated ?? true)
-    }
-}
-
-// SetupViewController
-extension UITabBarItem {
-    func setup(title: String, imageName: String, selectedImageName: String) {
-        self.title = title
-        self.image = UIImage(systemName: imageName)
-        self.selectedImage = UIImage(systemName: selectedImageName)
-    }
-}
