+++ Netflix/Protocols/Coordinator.swift
@@ -1,6 +1,5 @@
 import Foundation
 
 protocol Coordinator {
-    var childCoordinators: [Coordinator] { get }
     func start()
 }
